Rename OverrideStatementRewriterWith to UseStatementRewriter; rename oracle package to oracle12
