+++ oracle12/src/main/java/org/jdbi/v3/oracle12/OracleReturning.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.oracle;
+package org.jdbi.v3.oracle12;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
