+++ oracle12/src/test/java/org/jdbi/v3/oracle12/TestGetGeneratedKeysOracle.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jdbi.v3.oracle;
+package org.jdbi.v3.oracle12;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
